requirements/requirements-freeze.txt
absl-py==2.0.0
affine==2.4.0
aiohttp==3.8.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
albumentations==1.3.1
amqp==5.1.1
annotated-types==0.5.0
antlr4-python3-runtime==4.9.3
anyio==3.7.1
appdirs==1.4.4
arrow==1.2.3
async-timeout==4.0.3
asyncssh==2.13.2
atpublic==4.0
attrs==23.1.0
backoff==2.2.1
beautifulsoup4==4.12.2
billiard==4.1.0
black==23.9.1
blessed==1.20.0
bpytop==1.0.68
cachetools==5.3.1
celery==5.3.4
certifi==2022.12.7
cffi==1.15.1
cfgv==3.4.0
charset-normalizer==2.1.1
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cligj==0.7.2
cmake==3.25.0
colorama==0.4.6
configobj==5.0.8
contourpy==1.1.1
croniter==1.4.1
cryptography==41.0.4
cycler==0.11.0
dateutils==0.6.12
decorator==5.1.1
deepdiff==6.5.0
dictdiffer==0.9.0
diskcache==5.6.3
distlib==0.3.7
distro==1.8.0
dpath==2.1.6
dulwich==0.21.6
dvc==3.22.0
dvc-data==2.16.3
dvc-gs==2.22.1
dvc-http==2.30.2
dvc-objects==1.0.1
dvc-render==0.6.0
dvc-studio-client==0.15.0
dvc-task==0.3.0
fastapi==0.103.1
filelock==3.12.4
Fiona==1.9.4.post1
flatten-dict==0.4.2
flufl.lock==7.1.1
fonttools==4.42.1
frozenlist==1.4.0
fsspec==2023.9.1
funcy==2.0
gcsfs==2023.9.1
geopandas==0.14.0
gitdb==4.0.10
GitPython==3.1.37
google-api-core==2.11.1
google-auth==2.23.0
google-auth-oauthlib==1.0.0
google-cloud-core==2.3.3
google-cloud-storage==2.11.0
google-crc32c==1.5.0
google-resumable-media==2.6.0
googleapis-common-protos==1.60.0
gpustat==1.1.1
grandalf==0.8
grpcio==1.58.0
h11==0.14.0
hydra-core==1.3.2
identify==2.5.29
idna==3.4
imagededup==0.3.1
imageio==2.31.3
inquirer==3.1.3
iterative-telemetry==0.0.8
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
kiwisolver==1.4.5
kombu==5.3.2
lazy_loader==0.3
lightning==2.0.9
lightning-cloud==0.5.38
lightning-utilities==0.9.0
lit==15.0.7
Markdown==3.4.4
markdown-it-py==3.0.0
MarkupSafe==2.1.2
matplotlib==3.8.0
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.0.6
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.0
nodeenv==1.8.0
numpy==1.24.1
nvidia-ml-py==12.535.108
oauthlib==3.2.2
omegaconf==2.3.0
opencv-python==4.8.0.76
opencv-python-headless==4.8.0.76
ordered-set==4.1.0
orjson==3.9.7
packaging==23.1
pandas==2.1.1
pathspec==0.11.2
Pillow==9.3.0
platformdirs==3.10.0
pre-commit==3.4.0
prompt-toolkit==3.0.39
protobuf==4.24.3
psutil==5.9.5
pyarrow==13.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.1.1
pydantic_core==2.4.0
pydot==1.4.2
pygit2==1.13.0
Pygments==2.16.1
pygtrie==2.5.0
PyJWT==2.8.0
pyparsing==3.1.1
pyproj==3.6.1
python-dateutil==2.8.2
python-editor==1.0.4
python-multipart==0.0.6
pytorch-lightning==2.0.9
pytz==2023.3.post1
PyWavelets==1.4.1
PyYAML==6.0.1
qudida==0.0.4
rasterio==1.3.8
ray==2.7.0
readchar==4.0.5
referencing==0.30.2
requests==2.28.1
requests-oauthlib==1.3.1
rich==13.5.3
rpds-py==0.10.3
rsa==4.9
ruamel.yaml==0.17.32
ruamel.yaml.clib==0.2.7
scikit-image==0.21.0
scikit-learn==1.3.1
scipy==1.11.2
scmrepo==1.3.1
shapely==2.0.1
shortuuid==1.0.11
shtab==1.6.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
snuggs==1.4.7
soupsieve==2.5
sqltrie==0.7.0
starlette==0.27.0
starsessions==1.3.0
sympy==1.12
tabulate==0.9.0
tensorboard==2.14.0
tensorboard-data-server==0.7.1
tensorboardX==2.6.2.2
threadpoolctl==3.2.0
tifffile==2023.9.18
tomlkit==0.12.1
torch==2.0.1+cu118
torchmetrics==1.2.0
torchsummary==1.5.1
torchvision==0.15.2+cu118
tqdm==4.66.1
traitlets==5.10.0
triton==2.0.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==1.26.13
uvicorn==0.23.2
vine==5.0.0
virtualenv==20.24.5
voluptuous==0.13.1
wcwidth==0.2.6
websocket-client==1.6.3
websockets==11.0.3
Werkzeug==2.3.7
yarl==1.9.2
zc.lockfile==3.0.post1
